Move fetching of the sponsorblock segments out of the store
